requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
aniso8601==8.1.1
    # via flask-restful
bcrypt==3.2.0
    # via flask-bcrypt
blinker==1.4
    # via flask-mail
certifi==2020.12.5
    # via requests
cffi==1.14.4
    # via bcrypt
chardet==4.0.0
    # via requests
click==7.1.2
    # via flask
decorator==4.4.2
    # via validators
dnspython==2.1.0
    # via -r requirements.in
flask-bcrypt==0.7.1
    # via -r requirements.in
flask-cors==3.0.10
    # via -r requirements.in
flask-jwt-extended==3.25.0
    # via -r requirements.in
flask-mail==0.9.1
    # via -r requirements.in
flask-restful==0.3.8
    # via -r requirements.in
flask-script==2.0.6
    # via -r requirements.in
flask==1.1.2
    # via
    #   -r requirements.in
    #   flask-bcrypt
    #   flask-cors
    #   flask-jwt-extended
    #   flask-mail
    #   flask-restful
    #   flask-script
gunicorn==20.0.4
    # via -r requirements.in
idna==2.10
    # via requests
itsdangerous==1.1.0
    # via flask
jinja2==2.11.3
    # via flask
markupsafe==1.1.1
    # via jinja2
pycparser==2.20
    # via cffi
pyjwt==1.7.1
    # via flask-jwt-extended
pymongo==3.11.3
    # via -r requirements.in
python-dotenv==0.15.0
    # via -r requirements.in
pytz==2021.1
    # via flask-restful
requests==2.25.1
    # via -r requirements.in
six==1.15.0
    # via
    #   bcrypt
    #   flask-cors
    #   flask-jwt-extended
    #   flask-restful
    #   validators
urllib3==1.26.3
    # via requests
validators==0.18.2
    # via -r requirements.in
werkzeug==1.0.1
    # via
    #   flask
    #   flask-jwt-extended

# The following packages are considered to be unsafe in a requirements file:
# setuptools
